Temporary fix for #234
