+++ src/main/java/com/legacy/aether/Aether.java
@@ -1,6 +1,7 @@
 package com.legacy.aether;
 
 import net.minecraft.util.ResourceLocation;
+import net.minecraftforge.common.ForgeVersion;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
 import net.minecraftforge.fml.common.Mod.Instance;
@@ -12,6 +13,7 @@ import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import com.legacy.aether.advancements.AetherAdvancements;
 import com.legacy.aether.entities.AetherEntities;
 import com.legacy.aether.networking.AetherNetworkingManager;
+import com.legacy.aether.player.AetherAdvancementHandler;
 import com.legacy.aether.player.capability.PlayerAetherManager;
 import com.legacy.aether.registry.AetherRegistries;
 import com.legacy.aether.registry.AetherRegistryEvent;
@@ -58,7 +60,12 @@ public class Aether
 		AetherWorld.initialization();
 
 		ServerProxy.registerEvent(new AetherEventHandler());
-
+		
+		if(ForgeVersion.getBuildVersion() >= 2565)
+		{
+			ServerProxy.registerEvent(new AetherAdvancementHandler());
+		}
+		
 		proxy.initialization();
 	}
