+++ src/main/java/com/legacy/aether/player/PlayerAetherEvents.java
@@ -1,13 +1,15 @@
 package com.legacy.aether.player;
 
-import net.minecraft.advancements.Advancement;
+import com.legacy.aether.Aether;
+import com.legacy.aether.advancements.AetherAdvancements;
+import com.legacy.aether.items.ItemsAether;
+import com.legacy.aether.player.capability.PlayerAetherProvider;
+
 import net.minecraft.command.CommandBase;
 import net.minecraft.command.CommandClearInventory;
 import net.minecraft.entity.Entity;
-import net.minecraft.entity.item.EntityItem;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
-import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.event.AttachCapabilitiesEvent;
@@ -17,7 +19,6 @@ import net.minecraftforge.event.entity.living.LivingAttackEvent;
 import net.minecraftforge.event.entity.living.LivingDeathEvent;
 import net.minecraftforge.event.entity.living.LivingEvent.LivingUpdateEvent;
 import net.minecraftforge.event.entity.living.LivingHurtEvent;
-import net.minecraftforge.event.entity.player.AdvancementEvent;
 import net.minecraftforge.event.entity.player.PlayerEvent.BreakSpeed;
 import net.minecraftforge.event.entity.player.PlayerEvent.Clone;
 import net.minecraftforge.event.entity.player.PlayerEvent.Visibility;
@@ -26,13 +27,6 @@ import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.gameevent.PlayerEvent.PlayerChangedDimensionEvent;
 import net.minecraftforge.fml.common.gameevent.PlayerEvent.PlayerLoggedInEvent;
 import net.minecraftforge.fml.common.gameevent.PlayerEvent.PlayerRespawnEvent;
-
-import com.legacy.aether.Aether;
-import com.legacy.aether.advancements.AetherAdvancements;
-import com.legacy.aether.items.ItemsAether;
-import com.legacy.aether.networking.AetherNetworkingManager;
-import com.legacy.aether.networking.packets.PacketAchievement;
-import com.legacy.aether.player.capability.PlayerAetherProvider;
 
 public class PlayerAetherEvents
 {
@@ -229,39 +223,4 @@ public class PlayerAetherEvents
 			}
 		}
 	}
-
-	@SubscribeEvent
-	public void onAchievementGet(AdvancementEvent event)
-	{
-		Advancement advancement = event.getAdvancement();
-		EntityPlayer player = event.getEntityPlayer();
-
-		String unlocalizedName = advancement.getDisplayText().getUnformattedText();
-
-		if (!unlocalizedName.contains("aether_legacy"))
-		{
-			return;
-		}
-
-		int achievementType = unlocalizedName.contains("bronze_dungeon") ? 1 : unlocalizedName.contains("silver_dungeon") ? 2 : 0;
-
-		if (!player.world.isRemote)
-		{
-			if (unlocalizedName.contains("enter_aether"))
-			{
-				if (!player.inventory.addItemStackToInventory(new ItemStack(ItemsAether.lore_book)))
-				{
-					player.world.spawnEntity(new EntityItem(player.world, player.posX, player.posY, player.posZ, new ItemStack(ItemsAether.lore_book)));
-				}
-
-				if (!player.inventory.addItemStackToInventory(new ItemStack(ItemsAether.golden_parachute)))
-				{
-					player.world.spawnEntity(new EntityItem(player.world, player.posX, player.posY, player.posZ, new ItemStack(ItemsAether.golden_parachute)));
-				}
-			}
-
-			AetherNetworkingManager.sendTo(new PacketAchievement(achievementType), (EntityPlayerMP) player);
-		}
-	}
-
 }
